move create_jobs(ÃÃ) to JobBuilder class
