+++ src/osekit/public/project.py
@@ -26,7 +26,6 @@ from osekit.core.ltas_dataset import LTASDataset
 from osekit.core.spectro_dataset import SpectroDataset
 from osekit.public.transform import OutputType, Transform
 from osekit.utils.core import (
-    file_indexes_per_batch,
     get_umask,
     locked,
 )
@@ -626,11 +625,12 @@ class Project:
             "dataset-json-path": self.folder / "project.json",
         }
 
-        self.create_jobs(
+        self.job_builder.create_jobs(
             audio_dataset=ads,
             script_path=Path(export_transform.__file__),
             script_args=script_args,
             job_name=name,
+            output_folder=self.folder / self.SUBFOLDERS["log"],
             nb_jobs=nb_jobs,
         )
 
@@ -675,45 +675,6 @@ class Project:
         )
 
         return ads_json, sds_json
-
-    def create_jobs(
-        self,
-        audio_dataset: AudioDataset,
-        script_path: Path,
-        script_args: dict,
-        job_name: str,
-        nb_jobs: int = 1,
-    ) -> None:
-        """Create the jobs corresponding to each batch.
-
-        Parameters
-        ----------
-        audio_dataset: AudioDataset
-            The ``AudioDataset`` the transform is based on.
-        script_path: Path
-            Path to the export script.
-        script_args: dict
-            Arguments passed to the export script.
-        job_name: str
-            Name of the job.
-            If there are multiple batches, each batch will be suffixed
-            with "_{index}".
-        nb_jobs: int
-            Number of batches used to run the transform.
-            Each batch will run in a separate job.
-
-        """
-        batch_indexes = file_indexes_per_batch(
-            total_nb_files=len(audio_dataset.data),
-            nb_batches=nb_jobs,
-        )
-        for index, (start, stop) in enumerate(batch_indexes):
-            self.job_builder.create_job(
-                script_path=script_path,
-                script_args=script_args | {"first": start, "last": stop},
-                name=job_name + (f"_{index}" if len(batch_indexes) > 1 else ""),
-                output_folder=self.folder / self.SUBFOLDERS["log"],
-            )
 
     def _add_spectro_dataset(
         self,
