+++ src/osekit/utils/job.py
@@ -14,6 +14,9 @@ from typing import TYPE_CHECKING, Literal
 
 from pandas import Timedelta
 
+from osekit.core.audio_dataset import AudioDataset
+from osekit.utils.core import file_indexes_per_batch
+
 if TYPE_CHECKING:
     from pathlib import Path
 
@@ -97,7 +100,7 @@ class Job:
         """
         config = JobConfig() if config is None else config
         self.script_path = script_path
-        self.script_args = script_args if script_args else {}
+        self.script_args = script_args or {}
         self.nb_nodes = config.nb_nodes
         self.ncpus = config.ncpus
         self.ngpus = config.ngpus
@@ -534,6 +537,48 @@ class JobBuilder:
         """
         self.config = config
         self.jobs = []
+
+    def create_jobs(
+        self,
+        audio_dataset: AudioDataset,
+        script_path: Path,
+        script_args: dict,
+        output_folder: Path,
+        job_name: str = "osekit_transform",
+        nb_jobs: int = 1,
+    ) -> None:
+        """Create the jobs corresponding to each batch.
+
+        Parameters
+        ----------
+        audio_dataset: AudioDataset
+            The ``AudioDataset`` the transform is based on.
+        script_path: Path
+            Path to the export script.
+        script_args: dict
+            Arguments passed to the export script.
+        job_name: str
+            Name of the job.
+            If there are multiple batches, each batch will be suffixed
+            with "_{index}".
+        output_folder: Path
+            Folder in which the job output log files are saved.
+        nb_jobs: int
+            Number of batches used to run the transform.
+            Each batch will run in a separate job.
+
+        """
+        batch_indexes = file_indexes_per_batch(
+            total_nb_files=len(audio_dataset.data),
+            nb_batches=nb_jobs,
+        )
+        for index, (start, stop) in enumerate(batch_indexes):
+            self.create_job(
+                script_path=script_path,
+                script_args=script_args | {"first": start, "last": stop},
+                name=job_name + (f"_{index}" if len(batch_indexes) > 1 else ""),
+                output_folder=output_folder,
+            )
 
     def create_job(
         self,
